src/requirements-new.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.58.4
fsspec==2024.6.1
idna==3.10
importlib_metadata==8.7.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.3.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
open3d==0.19.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.16.0
setuptools==70.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
zipp==3.23.0
